requirements.txt
cryptography>=41.0.0
pytest>=7.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
